LayoutTests/editing/spelling/grammar-markers-expected.txt

This tests whether WebKit can render grammar markers when pasting text if it has a grammar-checker. To test manually, copy the text in the first div element and paste it to the second div element. This test succeeds when WebKit renders a grammar marker under a word "has".

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS hasMarker is true
PASS successfullyParsed is true

TEST COMPLETE
You has the right.
You has the right.
